requirements.txt
flask
psutil
mysql-connector
mysql-connector-python
bcrypt